requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2022.9.24
greenlet==2.0.1
psycopg2-binary==2.9.5
python-telegram-bot==13.14
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==1.4.44
sqlparse==0.4.3
tornado==6.1
tzdata==2022.6
tzlocal==4.2
python-dotenv==0.21.0
